+++ tests/integration/fixtures/__init__.py
@@ -216,7 +216,7 @@ class SentryLike:
         if dsn_key is None:
             dsn_key = self.get_dsn_public_key(project_id, dsn_key_idx)
 
-        url = f"/api/{project_id}/spans/?sentry_key={dsn_key}"
+        url = f"/api/{project_id}/traces-data/?sentry_key={dsn_key}"
 
         if json:
             headers = {
